refactor network status command to have the printing implementation in one place. Adapted podman network status command
