+++ pkg/network/network.go
@@ -7,7 +7,8 @@ import (
 )
 
 const MINIMUM_VERSION string = "1.5.0"
-const MINIMUM_VERSION_MESSAGE string = "Detected that the skupper version installed in the namespace is version %s. The CLI requires version %s. To update the installation, please follow the instructions found here: https://skupper.io/docs/index.html"
+const MINIMUM_PODMAN_VERSION string = "1.6.0"
+const MINIMUM_VERSION_MESSAGE string = "Version incompatibility:\n \tDetected that the skupper version installed in the namespace is version %s. The CLI requires at least version %s. To update the installation, please follow the instructions found here: https://skupper.io/install/index.html"
 
 type SkupperStatus struct {
 	NetworkStatus *NetworkStatusInfo
